Introduce ORM in the stack
For this project I thought that ORM was a waste of time. But for
presenting the data in a rest way and passing stuff to flask-restplus
marshalling module ORM it makes sense.

Just introducting ORM where it really need to be. Doesn't plan to
replace all core with orm. SQL Core is still available.